Approximate Pattern Matching Problem
We say that a k-mer Pattern appears as a substring of Text with at most d mismatches if there is some k-mer substring Pattern' of Text having d or fewer mismatches with Pattern; that is, HammingDistance(Pattern, Pattern') ≤ d. Our observation that a DnaA box may appear with slight variations leads to the following generalization of the Pattern Matching Problem.

Approximate Pattern Matching Problem:  Find all approximate occurrences of a pattern in a string.
 Input: Strings Pattern and Text along with an integer d.
 Output: All starting positions where Pattern appears as a substring of Text with at most d mismatches.

Code Challenge (3 points): Write a function ApproximatePatternMatching to solve the Approximate Pattern Matching Problem. (Make sure to use HammingDistance as a subroutine!)