Count(Motifs)
Now that we can generate Count(Motifs), in order to compute Profile(Motifs), we simply need to divide each element of the count matrix by the number of rows in the count matrix, which we leave to you as an exercise.

Code Challenge (3 points): Write a function Profile(Motifs) that takes Motifs as input and returns their profile matrix as a dictionary of lists. Then place this function into Motifs.py. Make sure that you use the Count(Motifs) function that we already wrote as a subroutine!